pyproject.toml
[project]
name = "py-flamegraph"
version = "0.1.0"
description = "Python profiling with Pyroscope + Kafka via eBPF (Grafana Alloy)"
readme = "README.md"
requires-python = ">=3.11,<3.13"
dependencies = [
    "confluent-kafka[avro,schema-registry]>=2.3.0",
    "python-dotenv>=1.0.0",
]

[build-system]
requires = ["uv_build>=0.9.2,<0.10.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "pytest>=8.4.2",
]
